+++ cucumber-expressions/ruby/spec/cucumber/cucumber_expressions/custom_transform_spec.rb
@@ -13,7 +13,7 @@ module Cucumber
       end
       ### [color-constructor]
 
-      def == (other)
+      def ==(other)
         other.is_a?(Color) && other.name == name
       end
     end